Move also new studyprogramme-stuff in backend into its own file
Again, why did I not do this already last week?
